+++ scripts/prepyrus/src/main.rs
@@ -4,6 +4,11 @@ use serde::Deserialize;
 use std::fs;
 use std::io::{self, BufReader, Read, Write};
 use std::path::Path;
+use utils::BiblatexUtils;
+use validation::Metadata;
+
+mod utils;
+mod validation;
 
 /// Prepyrus is a tool for verifying and processing MDX files
 /// that contain citations in Chicago author-date style.
@@ -47,7 +52,7 @@ fn main() {
     let all_entries = bibliography.into_vec();
 
     // Phase 1: Verify MDX files
-    verify_mdx_files(mdx_paths.clone(), &all_entries);
+    validation::verify_mdx_files(mdx_paths.clone(), &all_entries);
 
     // Phase 2: Process MDX files (requires arg[3] mode to be set to "process")
     if args[3].eq("process") {
@@ -78,56 +83,6 @@ fn verify_arguments(args: &Vec<String>) {
         eprintln!("Invalid mode. Please provide either 'verify' or 'process'.");
         std::process::exit(1);
     }
-}
-
-fn verify_mdx_files(mdx_paths: Vec<String>, all_entries: &Vec<Entry>) {
-    let mut article_count = 0;
-    for mdx_path in &mdx_paths {
-        let (metadata, markdown_content, _) = match read_mdx_file(&mdx_path) {
-            Ok(data) => data,
-            Err(err) => {
-                if err.kind() == io::ErrorKind::InvalidData {
-                    eprintln!("Invalid MDX data format: {}", err);
-                    std::process::exit(1);
-                } else {
-                    eprintln!("Unexpected error reading MDX file: {}", err);
-                    std::process::exit(1);
-                }
-            }
-        };
-        if !metadata.is_article {
-            continue;
-        }
-        let citations = extract_citations_from_markdown(&markdown_content);
-        match verify_citations_format(&citations) {
-            Ok(_) => {}
-            Err(err) => {
-                eprintln!("Error verifying citations: {} in {}", err, mdx_path);
-                std::process::exit(1);
-            }
-        };
-        let citations_set = create_citations_set(citations);
-        if citations_set.is_empty() {
-            continue;
-        }
-        match match_citations_to_bibliography(citations_set, &all_entries) {
-            Ok(data) => data,
-            Err(err) => {
-                eprintln!(
-                    "Error matching citations to bibliography: {} in {}",
-                    err, mdx_path
-                );
-                std::process::exit(1);
-            }
-        };
-        article_count += 1;
-    }
-    // todo : construct the array of structs of each article and return the full list
-    println!(
-        "===Integrity verification OK: {} files verified, including {} articles",
-        mdx_paths.len(),
-        article_count
-    );
 }
 
 // todo : arg should be a struct based on {path, metadata, markdown_content, full_file_content} as well as all_bib_entries
@@ -243,207 +198,11 @@ fn filter_mdx_paths_for_exceptions(mdx_paths: Vec<String>, exceptions: Vec<Strin
     filtered_paths
 }
 
-#[derive(Debug, Deserialize)]
-struct Metadata {
-    #[allow(dead_code)]
-    title: String,
-    #[allow(dead_code)]
-    description: String,
-    #[serde(rename = "isArticle")]
-    is_article: bool,
-    #[allow(dead_code)]
-    authors: Option<String>,
-    #[allow(dead_code)]
-    editors: Option<String>,
-    #[allow(dead_code)]
-    contributors: Option<String>,
-}
-
 fn write_html_to_mdx_file(path: &str, content: &str) -> io::Result<()> {
     let file = fs::File::create(path)?;
     let mut writer = io::BufWriter::new(file);
     writer.write_all(content.as_bytes())?;
     Ok(())
-}
-
-fn read_mdx_file(path: &str) -> io::Result<(Metadata, String, String)> {
-    let file = fs::File::open(path)?;
-    let mut reader = BufReader::new(file);
-    let mut content = String::new();
-    reader.read_to_string(&mut content)?;
-
-    // Extract metadata enclosed in `---` at the start of the file
-    let parts: Vec<&str> = content.splitn(3, "---").collect();
-    if parts.len() != 3 {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("Unable to extract metadata in {}", path),
-        ));
-    }
-
-    let metadata_str = parts[1];
-    let metadata: Metadata = match serde_yaml::from_str(metadata_str) {
-        Ok(data) => data,
-        Err(err) => {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                format!("{} in {}", err, path),
-            ))
-        }
-    };
-    let markdown_content = parts[2].to_string();
-    let full_file_content = content.clone();
-
-    Ok((metadata, markdown_content, full_file_content))
-}
-
-/// Extract citations from a markdown string
-/// The citations are assumed to be Chicago author-date style
-/// and in the format (Author 2021) or (Author 2021, 123)
-fn extract_citations_from_markdown(markdown: &String) -> Vec<String> {
-    //      Regex explanation
-    //
-    //      \(      Match an opening parenthesis
-    //      ([A-Z]  Match a capital letter
-    //      [^()]*? Match any character except opening and closing parenthesis
-    //      \d+     Match one or more digits
-    //      (?:     Start a non-capturing group
-    //      ,       Match a comma
-    //      [^)]*   Match any character except closing parenthesis
-    //      )?      End the non-capturing group and make it optional
-    //      \)      Match a closing parenthesis
-    //
-    // The regex will match citations in the format (Author 2021) or (Author 2021, 123)
-    //
-    let citation_regex = Regex::new(r"\(([A-Z][^()]*?\d+(?:,[^)]*)?)\)").unwrap();
-    let mut citations = Vec::new();
-
-    for line in markdown.lines() {
-        for captures in citation_regex.captures_iter(line) {
-            let citation = captures.get(1).unwrap().as_str().trim();
-            citations.push(citation.to_string());
-        }
-    }
-    citations
-}
-
-fn verify_citations_format(citations: &Vec<String>) -> Result<(), io::Error> {
-    for citation in citations {
-        let citation_split = citation.splitn(2, ',').collect::<Vec<&str>>();
-        let first_part = citation_split[0].trim();
-        let has_year = first_part.split_whitespace().any(|word| {
-            if let Ok(num) = word.parse::<u32>() {
-                num >= 1000 && num <= 9999
-            } else {
-                false
-            }
-        });
-        if !has_year {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                format!("Citation is malformed or is missing year: ({})", citation),
-            ));
-        }
-    }
-    Ok(())
-}
-
-fn create_citations_set(citations: Vec<String>) -> Vec<String> {
-    let mut citations_set = Vec::new();
-    for citation in citations {
-        let prepared_citation = citation
-            .splitn(2, ',')
-            .next()
-            .unwrap_or(&citation)
-            .to_string();
-        if !citations_set.contains(&prepared_citation) {
-            citations_set.push(prepared_citation);
-        }
-    }
-    citations_set
-}
-
-fn extract_year(date: &PermissiveType<Date>, reference: String) -> Result<i32, io::Error> {
-    match date {
-        PermissiveType::Typed(date) => match date.value {
-            DateValue::At(datetime) => Ok(datetime.year),
-            DateValue::After(datetime) => Ok(datetime.year),
-            DateValue::Before(datetime) => Ok(datetime.year),
-            DateValue::Between(start, _end) => Ok(start.year), // Or use end.year
-        },
-        _ => Err(io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("Unable to retrieve year for: {}", reference),
-        )),
-    }
-}
-
-/// Use this to extract from a Spanned<Chunk> vector
-/// ```
-/// let address = extract_spanned_chunk(&address_spanned);
-/// ```
-fn extract_spanned_chunk(spanned_chunk: &[Spanned<Chunk>]) -> String {
-    spanned_chunk
-        .iter()
-        .filter_map(|spanned_chunk| match spanned_chunk.v {
-            Chunk::Normal(ref s) => Some(s.clone()),
-            _ => None,
-        })
-        .collect()
-}
-
-fn extract_publisher(publisher_data: &Vec<Vec<Spanned<Chunk>>>) -> String {
-    publisher_data
-        .iter()
-        .flat_map(|inner_vec| {
-            inner_vec
-                .iter()
-                .filter_map(|spanned_chunk| match spanned_chunk.v {
-                    Chunk::Normal(ref s) => Some(s.clone()),
-                    _ => None,
-                })
-        })
-        .collect()
-}
-
-/// Matches citations to the inputted bibliography
-/// the matched list is returned with full bibliographical details.
-/// If any citation is not found in the bibliography, an error is returned.
-fn match_citations_to_bibliography(
-    citations: Vec<String>,
-    bibliography: &Vec<Entry>,
-) -> Result<Vec<Entry>, io::Error> {
-    let mut unmatched_citations = citations.clone();
-    let mut matched_citations = Vec::new();
-
-    for citation in citations {
-        for entry in bibliography {
-            let author = entry.author().unwrap();
-            let author_last_name = author[0].name.clone();
-
-            let date: biblatex::PermissiveType<biblatex::Date> = entry.date().unwrap();
-            let year = extract_year(&date, citation.clone()).unwrap();
-
-            let author_year = format!("{} {:?}", author_last_name, year);
-
-            if citation == author_year {
-                unmatched_citations.retain(|x| x != &citation);
-                matched_citations.push(entry.clone());
-            }
-        }
-    }
-
-    if unmatched_citations.len() > 0 {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!(
-                "Citations not found in the library: ({:?})",
-                unmatched_citations
-            ),
-        ));
-    }
-
-    Ok(matched_citations)
 }
 
 fn generate_mdx_bibliography(entries: Vec<Entry>) -> String {
@@ -457,13 +216,13 @@ fn generate_mdx_bibliography(entries: Vec<Entry>) -> String {
             EntryType::Book => {
                 let author = entry.author().unwrap();
                 let title_spanned: &[biblatex::Spanned<biblatex::Chunk>] = entry.title().unwrap();
-                let title = extract_spanned_chunk(title_spanned);
+                let title = BiblatexUtils::extract_spanned_chunk(title_spanned);
                 let publisher_spanned: Vec<Vec<Spanned<Chunk>>> = entry.publisher().unwrap();
-                let publisher = extract_publisher(&publisher_spanned);
+                let publisher = BiblatexUtils::extract_publisher(&publisher_spanned);
                 let address_spanned: &[Spanned<Chunk>] = entry.address().unwrap();
-                let address = extract_spanned_chunk(address_spanned);
+                let address = BiblatexUtils::extract_spanned_chunk(address_spanned);
                 let date = entry.date().unwrap();
-                let year = extract_year(&date, entry.key.clone()).unwrap();
+                let year = BiblatexUtils::extract_year(&date, entry.key.clone()).unwrap();
                 let translators = entry.translator().unwrap_or(Vec::new());
 
                 if author.len() > 2 {
